refactor(voice): shrink state-machine parent modules (#4240)
